Fix X86Relocator's handling of indirect jmp
Was very tempted to port X86Relocator to capstone (the plan is to
gradually remove the remaining dependencies on udis86). Unfortunately it
is a big undertaking to do this part, as it is used by the Stalker, which
means we need to plug in a memory allocator that only uses pre-allocated
memory. Sigh.

Fixes frida/frida-core#5.