Make gum_script_from_string() asynchronous
This is the first part of moving the Script API to be fully asynchronous,
meaning that we never do any V8 calls from our MainContext thread.

The motivation is to be able to support debugging, where V8 goes into
a blocking loop waiting for commands when handling breakpoints and the
likes of. As we rely on our MainContext for exchanging messages with V8,
we must make sure that V8 stays out of our MainContext thread.