Implement gum_try_mprotect and gum_clear_cache for QNX
gum_try_mprotect includes special handling for the case
where an area cannot be marked PROT_WRITE because it is
a shared memory mapping which was opened read-only. This
appears to be the case for all .so code sections mapped.

The special handling includes remapping the necessary
areas in order to allow PROT_WRITE.