Optimize enumeration of module exports in a remote process
We now use the exports trie to avoid transferring huge amounts of data
from the remote process.

This was a hotspot when attaching to processes, and was previously so
slow that it was reportedly causing timeouts for some users. Today I
experienced this myself, and upon investigating I discovered that this
enumeration was taking many seconds on my desktop.