Add a tiny libc module
This is needed when we're running on kernels that don't support RWX pages
and we'd like to hook a function residing on the same memory page as for
example memcpy in the libc. As we need to flip the memory page to RW,
memcpy() our new data, and then flip it back to RX, we cannot use for
example memcpy() while the libc page is RW (i.e. not executable).