Start unifying the Interceptor's trampolines
This will allow attaching to a replaced function, and vice versa, fixing
a long-standing issue where this resulted in undefined behavior. It will
also simplify the architecture-specific backends, as they only need to
generate one kind of trampoline. This lays the groundwork for additional
complexity in the backends to support hooking tiny functions, which has
become a priority for Android/ARM.