Add Deflector API for making long-distance jumps
This makes it possible to reach a far-away destination with a minimal amount
of instructions. Only implemented on Linux/ARM for now. Works by looking for
a nearby ELF header and inserting a trampoline in an unused part, and taking
care of routing calls from multiple callers based on their return address.

This will allow us to hook tiny functions, and functions where there
is an early CALL/BL that prevents us from overwriting beyond it.