Fix synchronization logic in the Duktape runtime
Turns out that Duktape does not support pausing JS execution to allow
other native threads to execute JS in the same heap. The only possible
workaround is using coroutines, but those have strict limitations on when
a yield is possible.

So to keep things simple we adapt our runtime to never yield, and switch
our mutex to a recursive one, so we can still support re-entrancy through
NativeFunction calls.