Make on_enter and on_leave optional
We also avoid trapping the return when possible. Beside avoiding a little
bit of overhead, it more importantly allows hooking at arbitrary points,
not just at the start of a function. This means you could attach a
listener with only on_enter provided, and use that to hook at arbitrary
points. An example use-case is measuring code coverage.