Add Socket.listen() and Socket.connect() (#172)
By leveraging `GSocketClient` and `GSocketListener`, we get to re-use our
existing bindings for `GInputStream` and `GOutputStream`, and only need to add
bindings for `GIOStream` and `GSocketConnection`.

One could argue that this could be implemented in pure JavaScript by using
`NativeFunction`, but considering how many platform considerations need to be
taken into account, and that we already have this functionality in GIO (which
Frida depends on for its own needs), we might as well expose these APIs.

This will allow a Frida-specific `net` module to easily be built on top of these
building blocks. Consequently the idea is making it easy to use off-the-shelf
network modules from npm, e.g. for serving HTTP requests directly from an agent.