Fix crash on close() during I/O in the V8 runtime
Instead of introducing API to cancel individual operations, and
burdening developers with logic to cancel all before calling close(), we
now take care of these nitty gritty details.

By introducing a cancellable per object, we simply cancel it before scheduling
the close operation. The scheduling happens in an idle handler, which means
we can count on the cancellation being out of the way by the time we actually
close the stream.

This also gets rid of some duplicated logic with some new helper APIs
that should come handy in the other modules too.