Realize if supported, remap otherwise (#278)
When realize could fail, avoid it entirely and directly remap the
segment. This works where code signing isn't an issue but sandboxing
breaks executable file mappings (like Electra on iOS 11.1.2).

This means we avoid the dangerous RX->RW->RX fallback on iOS, where
another thread might be executing code on the page that is temporarily
RW while Interceptor is modifying it.