Simplify the Android dynamic linker integration
The primitives we were using made for a brittle implementation, as it
turns out the deeper ones are inlined on some versions of Android.

So instead try a higher level approach where we call an entrypoint near
dlopen() with flags matching how the module was already loaded, and
caller address matching the module's existing parent module.