Fix gum_thread_try_get_ranges() on latest Apple OSes (#338)
In recent libpthread, e.g. iOS 12.x, the `stacksize` field became `stackbottom`.
As the rest of the fields we care about are still valid, we can simply query
pthread_get_stacksize_np() to determine the size. This way there's no need to
add complexity to figure out the flavor of the structure (at least for now).