Add CModule API powered by TinyCC
For implementing short callback functions, where the cost of context-
switching in/out of a JS VM is higher than the actual code to be run,
even in unoptimized form.

Up until now we have had to use the CodeWriter APIs to generate such
callbacks, or deploy a shared library to Module.load().

This allows us to write inline C in such situations, and also allows
threads to be created, full control over synchronization, and calling
functions without libffi overhead. (Great for fuzzing purposes.)