Make linker seeking future-proof on Android (#360)
Per frida/frida#1026 and frida/frida-gum#354 - I looked into how the
linker is meant to be established in Android going forward.

According to

    https://source.android.com/devices/architecture/modular-system/runtime#bionic-changes

Google intends to maintain both `/system/bin/linker` and
`/system/bin/linker64` as a symlink, pointing to what the real file
will be.

This means if we look for those files and they are symlinks, the linker
should be mapped to memory using the files' realpath. So for future
changes, looking to see what the symlink resolves as should make this
work in all future version (and keep its backwards compatibility).

Tested this on Android 5, 7.2, 8.0.1, 9 and 10.