Improve detection of _pthread_exit() on iOS
The function has only been stable for a couple of major releases, and
changed quite a bit prior to that.

Turns out pthread_exit() has proven very stable across releases, so it
is better to parse it looking for the first BL instruction, which is the
call to _pthread_exit().

Kudos to @mrmacete for the neat idea!