Fix encoding of CpuContext pointers on Darwin/arm64e
We should expose the raw values used by the system to avoid losing data
when marshaling back and forth between the native structure and ours.

In this way even an exception caused by an authentication failure will
be exposed in its original form, and a “do nothing”-handler won't
corrupt the values in such a case.

This is even more important on macOS/arm64e, where the accessors raise
an exception if authentication fails, in turn resulting in our exception
handler deadlocking.