Fix Stalker entry with misaligned stack on arm64
The code being instrumented may misalign the stack, so we need to be
prepared for our prolog code triggering an exception.

We could consider optimizing this down the road by having a pair of
“nop” instructions before/after the prolog/epilog, and writing a pair
of fixup instructions there to avoid a trap on every entry/exit.

Kudos to @jacobcxdev for reporting and helping brainstorm this tricky
issue.