Enumerate Darwin imports from chained fixups (#546)
In the latest arm64e executables there are no binding opcodes, while
usually the same information can be reconstructed by walking the
metadata indexed by the LC_DYLD_CHAINED_FIXUPS command and stored in
linkedit.

The problem in doing so on images which have already been loaded in
memory by dyld is that the actual chains are lost because the bitfields
in the slots have already been replaced by the actual pointers.

This change reconstructs the correct slots by scanning the pages of the
segments which are pointed to by the fixups metadata and looking for
pointers to resolved imports.

Co-authored-by: Ole André Vadla Ravnås <oleavr@gmail.com>