Lower dlmalloc spin-lock sleep duration on Windows
The default of 50 ms is prone to introduce significant delay as soon as
threads start competing for the lock.

Instead pick a duration of zero, which means that SleepEx() only causes
the thread to relinquish the remainder of its time slice. This is
consistent with dlmalloc's spin-lock behavior on UNIX.