Add minimal context to NativeCallback (#561)
With this change, the NativeCallback JS function is called with `this`
bound to a CallbackContext.

This is an object with two properties, `context` and `returnAddress`,
which, similarly to their counterparts found in InvocationContext for
the Interceptor case, provide respectively a (minimal) CPU context and
the address of the caller code.

The CPU context is minimal, in that it contains only what's needed to
generate accurate backtraces, and it's read-only.

The primary use case for this addition is to provide a means for
generating accurate meaningful backtraces from the context of swizzled
ObjC methods (by using the ObjC.implement() API).