gumjs: Add ESM support (#575)
Using a human-readable bundle format that looks like this:

    📦
    59 /agent/index.js
    71 /agent/lib/logger.js
    ✄
    import { log } from './lib/logger.js';
    log('Hello World!');
    ✄
    export function log(...message) {
      console.log('[LOG]', ...message);
    }

The header specifies the assets that will be following, and the size of
each, in bytes. We include a “scissors” character with a newline before
and after each, but this is only used to improve readability, it does
not influence asset boundaries in any way.

The first entry with a .js extension is used as the entrypoint. However,
to make life easier for e.g. our REPL and its “--load” functionality, we
do allow multiple entrypoints. This is achieved by tacking on “\n✄\n”
followed by the next list of assets and their contents. In this way one
can add code to be run before and/or after by prepending/appending
additional modules, but without requiring tools to include complex
parsing and reassembly logic. (Which gets quite difficult when source
maps are involved.)

Module resolution follows the same rules as Node.js, which is why the
file extension is included in imports. A given module can have an alias,
so that given:

    import styles from 'ansi-styles';

An alias can be encoded like this:

    5010 /node_modules/ansi-styles/index.js
    ↻ ansi-styles

We also support source maps, where e.g. “/example.js” may be accompanied
by “/example.map”. Inline source maps are ignored.