meson: Fix compatibility with libdwarf from distros
We will amend our SDK build recipe to put the headers in a subdirectory,
where they belong, and also generate a .pc. Since latest libunwind
upstream now supplies a .pc we will eventually be able to get rid of the
hairy non-.pc logic.

Kudos to @smx-smx for helping figure this one out.