darwin-mapper: Initialize TLV descriptors on load
Within a module's image, before loading, the TLV descriptors thunk
function is set to _tlv_bootstrap, which calls abort at runtime.
The key values are also set to 0 on non-dyld_shared_cache images.

We create a new pthread_key at runtime, and store it in each TLV
descriptor. Additionally, we fix up the thunk function to tlv_get_addr,
which returns a pointer to the variable's storage as
((void **) gsbase)[descriptor->key] + descriptor->offset.

We're not handling any non-zero keys in the image, as those are only
used within the dyld_shared_cache.

Dyld's tlv_get_addr function retrieves the thread local storage by
using a descriptor's key to index into the thread local storage held in
gsbase on x86 or TPIDRRO_EL0 on arm64 to retrieve a thread- and module-
specific data area to offset into.

If this offset-specific area pointer is NULL, it will be lazily
allocated by tlv_get_addr through a call to
RuntimeState::_instantiateTLVs. This allocates the area, using a size
stored in a Dyld-private vector _tlvInfos, and copies over intial data.
When a module is loaded through Dyld, it stores the necessary size along
with the key in _tlvInfos.

Since we don't have a way to insert our newly generated key into this
array through public (and hence stable) APIs, we pre-allocate and
initialize the per-thread area for the initial thread in the runtime.

Co-authored-by: Ole André Vadla Ravnås <oleavr@gmail.com>