tests: Fix finding library by address on arm64be
On some systems, the segments of ld.so may not be contiguous, this can
result in libraries (particularly the VDSO since it is small) being
loaded into the gap between segments. Thus when searching for a
library which has been loaded into the gap by its address, the ld.so
module may be returned instead (since technically both modules contain
the address as FRIDA defines modules only by their base and size).