darwin-module: Swap-hook LLDB image notifier (#1057)
Since on modern iOS devices the `lldb_image_notifier` function is only
one instruction long, Interceptor currently fails to attach it. When
this happens, the module registry is left non-functional.

This change tries first to swap the pointer in the dyld "all image
infos" structure with our own, so there's no need to use Interceptor
in order to get image load/unload events.